backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
PyMuPDF==1.24.10
Pillow==10.4.0
pytesseract==0.3.13
httpx==0.27.2
pydantic==2.8.2
openai==1.51.0